fix(xo-web/backup): smart mode preview should ignore `xo:no-bak` tags
Introduced by 87a9fbe23

Fixes https://xcp-ng.org/forum/post/69797
