+++ packages/xo-web/src/xo-app/backup/overview/tab-jobs.js
@@ -3,7 +3,6 @@ import _ from 'intl'
 import ActionButton from 'action-button'
 import addSubscriptions from 'add-subscriptions'
 import Button from 'button'
-import constructQueryString from 'construct-query-string'
 import Copiable from 'copiable'
 import CopyToClipboard from 'react-copy-to-clipboard'
 import decorate from 'apply-decorators'
@@ -21,6 +20,7 @@ import { get } from '@xen-orchestra/defined'
 import { groupBy, isEmpty, map, some } from 'lodash'
 import { injectState, provideState } from 'reaclette'
 import { Proxy } from 'render-xo-item'
+import { smartModeToComplexMatcher } from 'smartModeToComplexMatcher'
 import { withRouter } from 'react-router'
 import {
   cancelJob,
@@ -360,7 +360,7 @@ class JobsTable extends React.Component {
         handler: (job, { goTo }) =>
           goTo({
             pathname: '/home',
-            query: { t: 'VM', s: constructQueryString(job.vms) },
+            query: { t: 'VM', s: smartModeToComplexMatcher(job.vms).toString() },
           }),
         disabled: job => job.type !== 'backup',
         label: _('redirectToMatchingVms'),
